azure_email_service/requirements.txt
Flask
azure-communication-email
